feat: rename and move remaining components lying around
